# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "detail": {
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginEntrada": {
        "additionalProperties": false,
        "properties": {
          "pin": {
            "examples": [
              "1234"
            ],
            "maxLength": 8,
            "minLength": 4,
            "title": "Pin",
            "type": "string"
          },
          "usuario": {
            "examples": [
              "bombero1"
            ],
            "maxLength": 40,
            "minLength": 3,
            "title": "Usuario",
            "type": "string"
          }
        },
        "required": [
          "usuario",
          "pin"
        ],
        "title": "LoginEntrada",
        "type": "object"
      },
...
      "TokenSalida": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expira_en": {
            "description": "segundos de validez del token",
            "title": "Expira En",
            "type": "integer"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioSalida"
          }
        },
        "required": [
          "access_token",
          "expira_en",
          "usuario"
        ],
        "title": "TokenSalida",
        "type": "object"
      },
...
      "UsuarioSalida": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "rol": {
            "enum": [
              "BOMBERO",
              "ADMIN"
            ],
            "title": "Rol",
            "type": "string"
          },
          "usuario": {
            "title": "Usuario",
            "type": "string"
          }
        },
        "required": [
          "id",
          "usuario",
          "nombre",
          "rol"
        ],
        "title": "UsuarioSalida",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "token de POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema que reemplaza a Nexus. **Banco de pruebas: no envía nada a la DIAN.** Dinero y galones viajan como texto decimal.",
    "title": "Estación de servicio — banco de pruebas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "423": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Locked"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Ingreso con usuario y PIN",
        "tags": [
          "sesión"
        ]
      }
    },
    "/auth/yo": {
      "get": {
        "operationId": "yo_auth_yo_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioSalida"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Usuario de la sesión actual",
        "tags": [
          "sesión"
        ]
      }
    },
...
  }
}